Add support for 64-bit ids in idhash.
We intend to use this with transports where dynamic "port numbers"
might be 32-bits.  This would allow us to formulate a 64-bit number
representing a conversation, and be able to find that conversation
by the 64-bit value.

Note that the hashed values are probably not perfectly optimal, as
only the low order bits are particularly significant in the hash.
We might want to consider XOR'ing in the upper bits to address that.

+++ src/core/endpt.c
@@ -18,7 +18,7 @@ struct nni_ep {
 	nni_tran_ep   ep_ops;  // transport ops
 	nni_tran *    ep_tran; // transport pointer
 	void *        ep_data; // transport private
-	uint32_t      ep_id;   // endpoint id
+	uint64_t      ep_id;   // endpoint id
 	nni_list_node ep_node; // per socket list
 	nni_sock *    ep_sock;
 	char          ep_addr[NNG_MAXADDRLEN];
@@ -78,7 +78,7 @@ nni_ep_sys_fini(void)
 uint32_t
 nni_ep_id(nni_ep *ep)
 {
-	return (ep->ep_id);
+	return ((uint32_t) ep->ep_id);
 }
 
 static void
